src/pyproject.toml
[tool.poetry]
name = "university-rating-system"
version = "0.1.0"
description = "Rating system for university"
authors = ["Nikita Reznikov <nikita.reznikov.live@mail.ru>"]
license = "MIT"
repository = "https://github.com/rnv812/UniversityRatingSystem"
packages = [{include = "university_rating_system"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
